Memory leaks taken out and tested. Lots of thread, timer, concurrency issues solved. DSOL upgraded for this purpose to 3.00.12.
DSOL 3.00.12 has the following extra features for this:

Use of Context rationalized and standardized. 
ContextUtil: extended methods to be symmetrical (bind-unbind) and (initial-given) contexts.
The Replication caches the context. 
Caching a context is removed from the Simulator.
ContextUtil.destroySubContext method added to recursively remove subcontexts and bindings.
Added a cleanUp() method to the Simulator to explicitly kill the worker thread that takes care of execution.
Added a cleanUp() method to the WorkerThread to explicitly kill the worker thread.
StatusBar got a cancelTimer() method to dispose of the timer so the StatusBar can be garbage collected.

Context becomes: 
/experiment_hash_code/replication_number_or_id/animation/2D/renderables
/experiment_hash_code/replication_number_or_id/animation/3D/renderables
/experiment_hash_code/replication_number_or_id/statistics/graphs

Code for binding a statistic:
  Context context = ContextUtil.lookup(simulator.getReplication().getContext(), "/statistics");
  ContextUtil.bind(context, this);

Code for binding a renderable2D:
  Context context = ContextUtil.lookup(simulator.getReplication().getContext(), "/animation/2D");
  ContextUtil.bind(context, this);

Removing the entire tree of an experiment:
  experiment.removeFromContext();
or if you just have a simulator:
  simulator.getReplication().getExperiment().removeFromContext();