XMLNetworkLaneParser: has to give error when two lanes have the same name. Check also added for     node     lanetype     roadtype     gtu     gtumix     route     routemix     shortestroute     shortestroutemix
XMLNetworkLaneParser angle in degrees or radians.

XMLNetworkLaneParser avoid double generators. Generators were wrongly stored at at the RoadTypeTag instead of at the LinkTag. A few structures have to be moved and changed. Also done for: FILL tag and BLOCK tag to keep everything symmetrical.

XMLNetworkLaneParser give an offset for the Generator. Allow in the XML code to put the GENERATOR on a distance from the start or end of the lane, or as a fraction of the length. Same for the BLOCK tag, and later the SENSOR tag and SINK tag. Examples: 
POSITION="10 m"
POSITION="BEGIN"
POSITION="END"
POSITION="END-10m"
POSITION="50%"