The network used in the CircularRoadIMB demo was re-used when the simulation was restarted. This resulted in severe complaints about duplicate names for Links and Nodes.
It was hard to find a suitable location to create the Network AND be able to create the IMB communication channels.

My current solution is to wrap the OTSModelInterface in another OTSModelInterface (the class ModelWrapper in SimpleIMBAnimator) and, in the constructModel method of the wrapper set the IMB connector just before calling the constructModel method of the contained (wrapped) OTSModelInterface. This is ugly, but it appears to get the job done.