The GTUType does not have static definitions anymore, to allow alternative GTUTypes for different situations. An enum called GTUType.DEFAULTS has been added that indicates the names of the GTUTypes that are defined by default. GTUTypes are registerd in the Network, and can be retrieved from the network. A method called addDefaultGtuTypes() has been added to the network to instantiate all default types in a default way.
The LinkType does not have static definitions anymore, to allow alternative LinkTypes for different situations.
An enum called LinkType.DEFAULTS has been added that indicates the names of the LinkTypes that are defined by default.
LinkTypes are registerd in the Network, and can be retrieved from the network.
A method called addDefaultLinkTypes() has been added to the network to instantiate all default types in a default way.

An interface called 'Definitions' has been added that allows to retrieve the LinkTypes and GTUTypes defined in the
current context. Both the Network interface and the PerceivableContext interface inherit from Definitions, so both
the GTUs and the Nodes/Links have access to the definitions that exist in the current simulation context.

The constructor of the OTSNetwork (and OTSRoadNetwork) has been given an boolean whether to construct the defaults or not.

The LaneType does not have static definitions anymore, to allow alternative LaneTypes for different situations.
An enum called LaneType.DEFAULTS has been added that indicates the names of the LaneTypes that are defined by default.
LaneTypes are registerd in the RoadNetwork, and can be retrieved from the road network.
A method called addDefaultLaneTypes() has been added to the RoadNetwork to instantiate all default types in a default way.

In order to give specific definitions to the network for Roads, a RoadNetwork interface and an OTSRoadNetwork class have been added.
These hold for instance information about the LaneTypes.

An interface called 'RoadDefinitions' has been added that allows to retrieve the LaneTypes defined in the
current context. The RoadNetwork interface inherit from Definitions (and therefore the OTSRoadNetwork class implements
the interfaces), so both the LaneBasedGTUs and the Nodes/Links have access to the Lane definitions that exist in the 
current simulation context. As RoadDefinitions inherits from Definitions, LinkTypes and GTUTypes are also accessible.

All code and demos has been adapted to work with the new GTUType, LinkType, LaneType, Definitions, and RoadDefinitions.