DJUTILS and DSOL have been updated to contain TimedEvents and strongly typed events rather than events that can take any payload.
The latest version of DSOL has a much more strict state machine that prevents unclear states between stopping the simulator from the outside and actually stopping of the run thread of the simulator. This solves problems with the runUntil() methods.  

The OTSNetwork now has the simulator stored in the network, which can prevent all classes to have to store a pointer to the simulator in addition to storing a pointer to the network.