Wrote TraceVerifier to debug/proof race conditions in complex software.
Fixed a bunch of race conditions in caching methods in AbstractGTU.java and AbstractLaneBasedGTU2 (should probably also be done in AbstractLaneBasedGTU).

classPath of ots-road needs fixing through the pom file (not yet done).

Added extra (temporary?) code to CircularRoadDemo and CircularRoadSwing to help find the race conditions in that demo.

Changed payload of Lane.GTU_ADD_EVENT (no longer contains the GTU object). This has consequences for TrafficLightSensor (which now needs access to the Network) and RoadSampler.