Update to DSOL version 4.0.1
Most important changes as a result of this update concern the
simplification of the Time notion for the Simulator and related objects.
Instead of generics <AbsoluteTime, RelativeTime, TimeObjct>, only the
generic <RelativeTime> remains. This makes writing the code a lot
easier. Since OTS works with absolute Time for the simulation events and
internal time notation, the OTSSimulatorInterface has been adapted with
a number of easy access methods for using Time objects and requesting
the absolute simulation time. The OTSReplication contains the absolute
startTime, which is the "zero" value for the absolute time. This means
that absolute simulation time is
replication.getStartTimeAbs().plus(relative_simulation_time), and the
relative simulation time can always be retrieved from an absolute time
by subtracting replication.getStartTimeAbs(). To make this even easier,
the OTSSimulatorInterface has a method getStartTimeAbs() that retrieves
the absolute start time from the corresponding replication.