Destroy GTU when not registered on a lane
There is still a problem with registration. Fractional positions often
get above 1.0, resulting in the GTU not being registered on any lane.
When this happens, An error message is printed, and the GTU is removed
from the simulation. Unfortunately, this triggers a new error for the
animation that still has to be resolved.

As a result, all demo simulation run, but several of them have problems
with the removal of GTUs durin the simulation run.