Issue #60. pom.xml: resource management in Eclipse
The /src/main/resources folder was also used for the test source folders
of the code. The latest versions of Eclipse turned the output folder of
the resources into /target/test-classes as a result, causing resources
to be unavailable. Since we copy the resources manually with Maven
anyway, there is no problem removing the normal resources as part of
testing in the <resources> tag in the pom file. This enables the correct
export of runnable jar files again.