Issue #55. Workaround for schema and includes for Eclipse
These changes ensure that the Eclipse editor can find the definitions
and includes (through the <xi:fallback> tag). Eclipse runs and
executable jar files can find the includes through the <xi:include> tag,
and the ots-file through the provided <xsi:schemaLocation>.

The nice thing is that the Eclipse XML editor and validation work
properly, so hints and explanations are given in the Eclipse editor. The
key/keyrefs for included elements give an error during editing, but that
disappears as soon as the editor is closed.